+++ src/main/java/be/looorent/TokenConfiguration.java
@@ -1,5 +1,6 @@
 package be.looorent;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import static java.lang.Math.min;
@@ -12,20 +13,22 @@ import static org.keycloak.common.util.Time.currentTime;
 public class TokenConfiguration {
 
     @JsonProperty("tokenLifespanInSeconds")
-    private Integer tokenLifespanInSeconds;
+    private final Integer tokenLifespanInSeconds;
+
+    @JsonCreator
+    public TokenConfiguration(@JsonProperty("tokenLifespanInSeconds") Integer tokenLifespanInSeconds) {
+        this.tokenLifespanInSeconds = tokenLifespanInSeconds;
+    }
 
     public Integer getTokenLifespanInSeconds() {
         return tokenLifespanInSeconds;
     }
 
-    public void setTokenLifespanInSeconds(Integer tokenLifespanInSeconds) {
-        this.tokenLifespanInSeconds = tokenLifespanInSeconds;
-    }
-
-    public int computeTokenExpiration(int maxExpiration, boolean longLivedTokenAllowed) {
-        return ofNullable(tokenLifespanInSeconds)
+    public int computeTokenExpiration(Long maxExpiration, boolean longLivedTokenAllowed) {
+        int furthestExpiration = ofNullable(maxExpiration).map(Long::intValue).orElse(0);
+        return ofNullable(getTokenLifespanInSeconds())
                 .map(lifespan -> currentTime() + lifespan)
-                .map(requestedExpiration -> longLivedTokenAllowed ? requestedExpiration : min(maxExpiration, requestedExpiration))
-                .orElse(maxExpiration);
+                .map(requestedExpiration -> longLivedTokenAllowed ? requestedExpiration : min(furthestExpiration, requestedExpiration))
+                .orElse(furthestExpiration);
     }
 }